+++ src/decisions/admin/publication.controller.ts
@@ -34,7 +34,11 @@ export class PublicationController {
   @Get('/:professionId/:organisationId/:year/publish')
   @Permissions(UserPermission.PublishDecisionData)
   @Render('admin/decisions/publication/new')
-  @BackLink('/admin/decisions/:professionId/:organisationId/:year')
+  @BackLink((request) =>
+    request.query.fromEdit === 'true'
+      ? '/admin/decisions/:professionId/:organisationId/:year/edit'
+      : '/admin/decisions/:professionId/:organisationId/:year',
+  )
   async new(
     @Param('professionId') professionId: string,
     @Param('organisationId') organisationId: string,